requirements.txt
# FastAPI and related packages
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic[email]>=2.0.0  # For email validation
pydantic-settings>=2.0.0
python-multipart>=0.0.6  # For handling file uploads
python-dotenv>=1.0.0  # For environment variables
aiofiles>=23.1.0  # For async file operations

# Supabase
supabase>=0.7.1
postgrest>=0.10.6

# OpenRouter client
httpx>=0.24.0  # Async HTTP client
cachetools>=5.3.0  # For caching

# Authentication and security
python-jose[cryptography]>=3.3.0  # For JWT handling
passlib[bcrypt]>=1.7.4  # For password hashing
pyjwt>=2.6.0  # For JWT handling

# Logging and monitoring
loguru>=0.6.0
sentry-sdk>=1.19.1

# Image processing
pillow>=9.5.0  # For image processing

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0  # For test coverage

# Development tools
black>=23.3.0  # Code formatting
isort>=5.12.0  # Import sorting
flake8>=6.0.0  # Linting
mypy>=1.2.0  # Type checking
